convert pk from Integer to Long

+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/test/java/org/dcm4chex/archive/ejb/session/StorageBeanTest.java
@@ -123,7 +123,7 @@ public class StorageBeanTest extends TestCase {
         ds.putAE(PrivateTags.CallingAET, CALLING_AET);
         ds.putAE(PrivateTags.CalledAET, CALLED_AET);
         ds.putAE(Tags.RetrieveAET, RETRIEVE_AET);
-        Integer assocpk = storage.initAssociation(CALLING_AET, CALLED_AET, RETRIEVE_AET);
+        Long assocpk = storage.initAssociation(CALLING_AET, CALLED_AET, RETRIEVE_AET);
         storage.store(assocpk, ds, "/", path.substring(1), (int) file.length(),
                 md.digest());
         storage.removeAssociation(assocpk);
